% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_data.R
\name{cluster_data}
\alias{cluster_data}
\title{cluster_data

is used to cluster all the samples in the study using FlowSOM clustering
and metaclustering into with ConsensusClusterPlus.}
\usage{
cluster_data(daf, seed, markers_to_use, clusters_nb)
}
\arguments{
\item{daf}{data single cell experiment}

\item{seed}{seed used to reproduce the results}

\item{markers_to_use}{markers to use for clustering}

\item{clusters_nb}{number of clusters}
}
\value{
daf data single cell experiment after clustering
}
\description{
cluster_data

is used to cluster all the samples in the study using FlowSOM clustering
and metaclustering into with ConsensusClusterPlus.
}
\examples{
cluster_data(daf,w234,data$lineage_markers,20)

}
