% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytofRUV-package.R
\docType{package}
\name{CytofRUV}
\alias{CytofRUV}
\title{CytofRUV}
\description{
CytofRUV CytofRUV is a computational algorithm which permits the integration
of data across CyTOF batches.
}
\details{
CytofRUV is able to load and normalise CyTOF datasets accross batches and also offers 
visualisation through diagnostic plots using the R-Shiny application before and
after CytofRUV normalisation.
}
