% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load_data

is used to load the fcs files from all the samples in the study,
the metadata file containing the details of each sample
and a panel file containing the details of all proteins used in the study.}
\usage{
load_data(wd_data, metadata_filename, panel_filename, cofactor = 5)
}
\arguments{
\item{wd_data}{Path to the directory containing all raw fcs files, the metadata file
and the panel file}

\item{panel_filename}{Panel filename containing the details of each marker}

\item{cofactor}{Cofactor for asinh transformation, default is 5 and set to NULL for untransformed data}

\item{metadata_name}{Metadata filename containing the details of each sample}
}
\value{
Datasets before normalisation
}
\description{
load_data

is used to load the fcs files from all the samples in the study,
the metadata file containing the details of each sample
and a panel file containing the details of all proteins used in the study.
}
