% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_data.R
\name{normalise_data}
\alias{normalise_data}
\title{normalise_data

contains all the functions to normalise several Cytof datasets by removing the
unwanted variation between datasets arising from experimental artefacts.}
\usage{
normalise_data(
  data,
  raw_data,
  rep_samples,
  norm_clusters,
  k,
  num_clusters,
  wd_data,
  dir_norm_data
)
}
\arguments{
\item{raw_data}{Raw data}

\item{rep_samples}{Replicated samples}

\item{norm_clusters}{Clusters to be normalised}

\item{k}{Dimension of the unwanted variation}

\item{num_clusters}{Total number of clusters}

\item{wd_data}{Path to the directory containing all raw fcs files, the metadata file
and the panel file}

\item{dir_norm_data}{Directory name containing all norm fcs files, the metadata file
and the panel file}

\item{daf}{Datasets before normalisation}
}
\value{
Normalised metadata file
}
\description{
normalise_data

contains all the functions to normalise several Cytof datasets by removing the
unwanted variation between datasets arising from experimental artefacts.
}
