% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEMD.R
\name{plotEMD}
\alias{plotEMD}
\title{plotEMD

Insert description of plotEMD here}
\usage{
plotEMD(
  daf,
  raw_dir,
  method_names = c("cytofRUV"),
  data_dir,
  k = NULL,
  filter_patients,
  x_axis_name,
  cols = c(raw = "black", batchadj = "indianred4", cytonorm = "darkmagenta",
    cytofruv_k1 = "#084081", cytofruv_k3 = "#0868AC", cytofruv_k5 = "#2B8CBE",
    cytofruv_k7 = "#4EB3D3", cytofruv_k10 = "#7BCCC4", cytofruv_k12 = "#A8DDB5",
    cytofruv_k15 = "darkgreen", cytofruv_k20 = "seagreen", cytofruv_k25 = "yellowgreen")
)
}
\arguments{
\item{daf}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}, this was generated after running examples present
in the vignette.}

\item{raw_dir}{String: Directory to file containing raw data.}

\item{method_names}{Vector: Strings that describe method names.
e.g. c(cytonorm, batchadj, cytofruv)}

\item{data_dir}{Vector: Strings of directories that contain appropriate data.
IMPORTANT: Order of directory strings should correlate with order of method names in method_names.
Note the directory for cytofRUV processed data is a vector of strings.
e.g. c("../Cytofnorm_Cytof_Package/EMD_metric_comp_norm_all_samples_all_cells.rds",
       "../BatchAdjust/BatchAdjust_95p_and_DG23_new_trial/EMD_metric_comp_norm_all_samples_all_cells.rds"
       c("CytofRUV_Norm_data_CLL2_HC1_all_cl_20_k5/EMD_metric_comp_norm_all_samples_all_cells.rds",
         "CytofRUV_Norm_data_CLL2_HC1_all_cl_20_k15/EMD_metric_comp_norm_all_samples_all_cells.rds",
         "CytofRUV_Norm_data_CLL2_HC1_all_cl_20_k20/EMD_metric_comp_norm_all_samples_all_cells.rds"
       )
       )}

\item{k}{Vector: cluster numbers to be inspected, default val: c(5,10,15)}

\item{filter_patients}{Vector: relevant patient IDs  e.g. val = c("LL1_B1","LL2_B1","LL3_B1") that is,
display patient IDs that contain: "LL1_B1","LL2_B1","LL3_B1".}

\item{x_axis_name}{String: Will be shared by all x-axis labels will share in common. E.g. "CLL", resulting
x-axis items = "CLL1, CLL2, CLL3"}

\item{cols}{Vector: Strings of colours for plot annotations. Default val:
c("raw"="black","batchadj"="indianred4","cytonorm"="darkmagenta","cytofruv_k1"="#084081","cytofruv_k3"="#0868AC","cytofruv_k5"="#2B8CBE","cytofruv_k7"="#4EB3D3","cytofruv_k10"="#7BCCC4","cytofruv_k12"= "#A8DDB5","cytofruv_k15"="darkgreen","cytofruv_k20"="seagreen","cytofruv_k25"="yellowgreen")}
}
\description{
plotEMD

Insert description of plotEMD here
}
