% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_Shiny.R
\name{launch_Shiny}
\alias{launch_Shiny}
\title{Launch Shiny

is used to  launch the CytofRUV Shiny app. It examines any batch effects present when comparing CyTOF data
from samples replicated across batches using four different diagnostics plots:
Median Protein Expression, Protein Expression Distributions, Clustering Results
and Cluster Proportions.}
\usage{
launch_Shiny(subset_percentage = 50)
launch_Shiny(daf)
}
\arguments{
\item{subset_percentage}{Percentage of data that exploratory plots shall display, default is 50%.}
}
\arguments{
\item{daf}{Dataset that the app's plots are based upon.}
}
\value{
Opens a browser window with an interactive Shiny application.
}
\description{
Launch Shiny

is used to  launch the CytofRUV Shiny app. It examines any batch effects present when comparing CyTOF data
from samples replicated across batches using four different diagnostics plots:
Median Protein Expression, Protein Expression Distributions, Clustering Results
and Cluster Proportions.
}
